+++ js/react/es6/JsonData.js
@@ -74,10 +74,10 @@
  * from the web service.
  *
  * [onViewUpdated] is an optional callback that can be used to handle custom DOM
- * updates after data has been fetched and state has been set. It would not be common
- * to use in most React Apps however if you want to execute standard JS to manipulate
- * or read DOM once data is ready then this property can be used. An example of this
- * is shown in [DataFormsJS\examples\log-table-react.htm]
+ * updates or Unit Testing during data state changes [isLoading, isLoaded, hasError].
+ * It would not be common to use in most React Apps however if you want to execute
+ * standard JS to manipulate or read DOM once data is ready then this property can be used.
+ * An example of this is shown in [DataFormsJS\examples\log-table-react.htm]
  *
  * [fetchOptions] and [fetchHeaders] allow for the app to control the request options
  * and send request headers for the component. The default options used are:
@@ -186,7 +186,7 @@ function HasError(props) {
  */
 function IsLoaded(props) {
     const show = (props.fetchState === 1);
-    if (!show) {
+    if (!show || !props.children) {
         return null;
     }
     return React.cloneElement(props.children, {
@@ -294,6 +294,10 @@ export default class JsonData extends React.Component {
         this.setState({
             fetchState: 0,
         }, () => {
+            // Allow Custom JavaScript events to run if defined
+            this.updateView();
+
+            // Fetch the data
             fetch(url, options)
             .then(response => {
                 const status = response.status;
@@ -353,7 +357,11 @@ export default class JsonData extends React.Component {
 
     updateView() {
         if (typeof this.props.onViewUpdated === 'function') {
-            this.props.onViewUpdated();
+            try {
+                this.props.onViewUpdated();
+            } catch (e) {
+                console.error(e);
+            }
         }
     }